+++ src/FunFair.Test.Common/IntegrationTestBase.cs
@@ -26,7 +26,7 @@ public abstract class IntegrationTestBase : LoggingTestBase
     /// </summary>
     /// <param name="output">XUnit output</param>
     /// <param name="dependencyInjectionRegistration">Registers services with dependency injection services.</param>
-    protected IntegrationTestBase(ITestOutputHelper output, Action<IServiceCollection> dependencyInjectionRegistration)
+    protected IntegrationTestBase(ITestOutputHelper output, Func<IServiceCollection, IServiceCollection> dependencyInjectionRegistration)
         : base(output: output, dependencyInjectionRegistration: dependencyInjectionRegistration)
     {
     }
@@ -38,7 +38,7 @@ public abstract class IntegrationTestBase : LoggingTestBase
     /// <param name="dependencyInjectionRegistration">Registers services with dependency injection services.</param>
     /// <param name="initializeServices">Initialises services.</param>
     [SuppressMessage(category: "ReSharper", checkId: "UnusedMember.Global", Justification = "May be used in derived test classes")]
-    protected IntegrationTestBase(ITestOutputHelper output, Action<IServiceCollection> dependencyInjectionRegistration, Action<IServiceProvider> initializeServices)
+    protected IntegrationTestBase(ITestOutputHelper output, Func<IServiceCollection, IServiceCollection> dependencyInjectionRegistration, Action<IServiceProvider> initializeServices)
         : base(output: output, dependencyInjectionRegistration: dependencyInjectionRegistration, initializeServices: initializeServices)
     {
     }