+++ CHANGELOG.md
@@ -8,10 +8,9 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Added
 ### Fixed
 ### Changed
-- FF-1429 - Updated SonarAnalyzer.CSharp to 8.36.0.43782
-- FF-1429 - Updated Meziantou.Analyzer to 1.0.696
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.36.1.44192
 - FF-1429 - Updated Meziantou.Analyzer to 1.0.697
+- Made Configure services return IServiceCollection.
 ### Removed
 
 <!--